КОНТРОЛЛЕР ЛАЙКОВ

+++ BeerParty.Data/Entities/ProfileInterection.cs
@@ -10,9 +10,9 @@ namespace BeerParty.Data.Entities
 {
     public class ProfileInteraction
     {
-        public int Id { get; set; }
-        public int FromProfileId { get; set; } // ID профиля, который ставит лайк
-        public int ToProfileId { get; set; }   // ID профиля, который получает лайк
+        public long Id { get; set; }
+        public long FromProfileId { get; set; } // ID профиля, который ставит лайк
+        public long ToProfileId { get; set; }   // ID профиля, который получает лайк
         public InteractionType Type { get; set; } // Лайк или дизлайк
         public DateTime Timestamp { get; set; }
